+++ src/main/java/com/clouway/core/TransactionExecutor.java
@@ -1,10 +1,12 @@
 package com.clouway.core;
 
+import java.math.BigDecimal;
+
 /**
  * Created byivan.genchev1989@gmail.com.
  */
 public interface TransactionExecutor {
 
-  Transaction execute(Double amount,String query);
+  Transaction execute(BigDecimal amount,String query);
 
 }
